feat(board): refactored more things into board settings: board slug and removed unused components
