+++ apps/web/src/views/board/components/BoardDropdown.tsx
@@ -2,14 +2,14 @@ import { useRouter } from "next/router";
 import { t } from "@lingui/core/macro";
 import {
   HiEllipsisHorizontal,
-  HiLink,
   HiOutlineCog6Tooth,
   HiOutlineDocumentDuplicate,
-  HiOutlineTrash,
   HiOutlineStar,
+  HiOutlineTrash,
   HiStar,
 } from "react-icons/hi2";
 import { IoArchiveOutline } from "react-icons/io5";
+
 import Dropdown from "~/components/Dropdown";
 import { usePermissions } from "~/hooks/usePermissions";
 import { useModal } from "~/providers/modal";
@@ -44,7 +44,9 @@ export default function BoardDropdown({
       void utils.board.byId.invalidate();
       if (variables.isArchived !== undefined) {
         showPopup({
-          header: variables.isArchived ? t`Board archived` : t`Board unarchived`,
+          header: variables.isArchived
+            ? t`Board archived`
+            : t`Board unarchived`,
           message: variables.isArchived
             ? t`The board has been archived.`
             : t`The board has been unarchived.`,
@@ -91,44 +93,39 @@ export default function BoardDropdown({
   const items = [
     ...(isTemplate && canCreateBoard
       ? [
-        {
-          label: t`Make template`,
-          action: () => openModal("CREATE_TEMPLATE"),
-          icon: (
-            <HiOutlineDocumentDuplicate className="h-[16px] w-[16px] text-dark-900" />
-          ),
-        },
-      ]
+          {
+            label: t`Make template`,
+            action: () => openModal("CREATE_TEMPLATE"),
+            icon: (
+              <HiOutlineDocumentDuplicate className="h-[16px] w-[16px] text-dark-900" />
+            ),
+          },
+        ]
       : []),
     ...(!isTemplate && canEditBoard
       ? [
-        {
-          label: t`Board settings`,
-          action: () => openModal("BOARD_SETTINGS"),
-          icon: <HiOutlineCog6Tooth className="h-[16px] w-[16px] text-dark-900" />,
-        },
-        {
-          label: t`Edit board URL`,
-          action: () => openModal("UPDATE_BOARD_SLUG"),
-          icon: <HiLink className="h-[16px] w-[16px] text-dark-900" />,
-        },
-      ]
+          {
+            label: t`Board settings`,
+            action: () => openModal("BOARD_SETTINGS"),
+            icon: (
+              <HiOutlineCog6Tooth className="h-[16px] w-[16px] text-dark-900" />
+            ),
+          },
+        ]
       : []),
     ...(!isTemplate && canArchiveBoard
       ? [
-        {
-          label: isArchived ? t`Unarchive board` : t`Archive board`,
-          action: handleArchiveOrUnarchive,
-          icon: (
-            <IoArchiveOutline className="h-[16px] w-[16px] text-dark-900" />
-          ),
-        },
-      ]
+          {
+            label: isArchived ? t`Unarchive board` : t`Archive board`,
+            action: handleArchiveOrUnarchive,
+            icon: (
+              <IoArchiveOutline className="h-[16px] w-[16px] text-dark-900" />
+            ),
+          },
+        ]
       : []),
     {
-      label: isFavorite
-        ? t`Remove from favorites`
-        : t`Add to favorites`,
+      label: isFavorite ? t`Remove from favorites` : t`Add to favorites`,
       action: handleToggleFavorite,
       icon: isFavorite ? (
         <HiStar className="h-[16px] w-[16px] text-dark-900" />
@@ -138,14 +135,14 @@ export default function BoardDropdown({
     },
     ...(canDeleteBoard
       ? [
-        {
-          label: isTemplate ? t`Delete template` : t`Delete board`,
-          action: () => openModal("DELETE_BOARD"),
-          icon: (
-            <HiOutlineTrash className="h-[16px] w-[16px] text-dark-900" />
-          ),
-        },
-      ]
+          {
+            label: isTemplate ? t`Delete template` : t`Delete board`,
+            action: () => openModal("DELETE_BOARD"),
+            icon: (
+              <HiOutlineTrash className="h-[16px] w-[16px] text-dark-900" />
+            ),
+          },
+        ]
       : []),
   ];
 
@@ -154,10 +151,7 @@ export default function BoardDropdown({
   }
 
   return (
-    <Dropdown
-      disabled={isLoading || isArchiveActionPending}
-      items={items}
-    >
+    <Dropdown disabled={isLoading || isArchiveActionPending} items={items}>
       <HiEllipsisHorizontal className="h-5 w-5 text-dark-900" />
     </Dropdown>
   );
